+++ src/main/java/com/tableManager/tableManager/controllers/TicketController.java
@@ -2,6 +2,7 @@ package com.tableManager.tableManager.controllers;
 
 import com.tableManager.tableManager.dto.TicketDTO;
 import com.tableManager.tableManager.model.Ticket;
+import com.tableManager.tableManager.model.User;
 import com.tableManager.tableManager.service.TicketService;
 import com.tableManager.tableManager.service.UserService;
 import org.springframework.http.HttpStatus;
@@ -26,15 +27,32 @@ public class TicketController {
 
     @PutMapping("/toggleTicket/{id}")
     public ResponseEntity<TicketDTO> toggleTicket(@PathVariable Long id) {
-        ticketService.setActive(id);
 
-        return new ResponseEntity<>(HttpStatus.OK);
+        Long currentUserId = userService.getCurrentUserId();
+        User user = userService.findbyId(currentUserId);
+
+        if(user.isEnabled()) {
+            ticketService.setActive(id);
+
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PutMapping("/answerTicket/{id}")
     public ResponseEntity<TicketDTO> answerTicket(@PathVariable Long id, @RequestBody TicketDTO ticketDTO) {
-        ticketService.answerTicket(id, ticketDTO);
-        return new ResponseEntity<>(HttpStatus.OK);
+
+        Long currentUserId = userService.getCurrentUserId();
+        User user = userService.findbyId(currentUserId);
+
+        if(user.isEnabled()) {
+            ticketService.answerTicket(id, ticketDTO);
+            return new ResponseEntity<>(HttpStatus.OK);
+
+        }else{
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PreAuthorize("hasAuthority('ADMIN')")
@@ -46,12 +64,13 @@ public class TicketController {
     @GetMapping("/user/{userId}")
     public ResponseEntity<?> getTicketsByUserId(@PathVariable Long userId) {
         Long currentUserId = userService.getCurrentUserId();
+        User user = userService.findbyId(currentUserId);
 
-        if(currentUserId.equals(userId)) {
+        if(currentUserId.equals(userId) && user.isEnabled()) {
 
             return ResponseEntity.ok(ticketService.findByUserId(userId));
         }else{
-            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
     }
 
@@ -60,9 +79,10 @@ public class TicketController {
         Long currentUserId = userService.getCurrentUserId();
         Long adminId = 1L;
         Ticket ticket = ticketService.findById(id);
+        User user = userService.findbyId(currentUserId);
 
         if(ticket != null && Objects.equals(ticket.getUser().getId(), currentUserId)
-            || ticket != null && Objects.equals(currentUserId, adminId)){
+            || ticket != null && Objects.equals(currentUserId, adminId) && user.isEnabled()){
 
             return new ResponseEntity<>(ticket, HttpStatus.OK);
         }else{
@@ -72,17 +92,25 @@ public class TicketController {
     @PostMapping("/newTicket")
     public ResponseEntity<Ticket> createTicket(@RequestBody TicketDTO ticketDTO) {
         Long currentUserId = userService.getCurrentUserId();
-        Ticket ticket = ticketService.createTicket(ticketDTO, currentUserId);
-        return new ResponseEntity<>(ticket, HttpStatus.CREATED);
+        User user = userService.findbyId(currentUserId);
+
+        if(user.isEnabled()) {
+            Ticket ticket = ticketService.createTicket(ticketDTO, currentUserId);
+            return new ResponseEntity<>(ticket, HttpStatus.CREATED);
+        }else{
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
     }
 
     @PutMapping("/update/{id}")
     public ResponseEntity<Ticket> updateTicket(@PathVariable Long id, @RequestBody TicketDTO ticketDTO) {
         Long currentUserId = userService.getCurrentUserId();
+        User user = userService.findbyId(currentUserId);
 
-        Ticket ticket = ticketService.updateTicket(id, ticketDTO, currentUserId);
 
-        if(ticket != null) {
+            Ticket ticket = ticketService.updateTicket(id, ticketDTO, currentUserId);
+
+        if(ticket != null && user.isEnabled()) {
             return new ResponseEntity<>(ticket, HttpStatus.OK);
         }else{
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
@@ -92,8 +120,11 @@ public class TicketController {
     @DeleteMapping("/delete/{id}")
     public void deleteTicket(@PathVariable Long id) {
         Long currentUserId = userService.getCurrentUserId();
+        User user = userService.findbyId(currentUserId);
 
-        ticketService.deleteById(id, currentUserId);
+        if(user.isEnabled()) {
+            ticketService.deleteById(id, currentUserId);
+        }
 
     }
 }
